# Conflicts:
#	docs/features/database.md

+++ bootui-core/src/main/java/io/github/jdubois/bootui/core/dto/RequestProfileTimingDto.java
@@ -4,8 +4,9 @@ package io.github.jdubois.bootui.core.dto;
  * Coarse timing breakdown for a profiled request.
  *
  * @param totalMs total request wall-clock time in milliseconds, or {@code null} when unknown
- * @param sqlMs summed duration of the correlated SQL statements in milliseconds
+ * @param sqlMs summed duration of the correlated SQL statements, in fractional milliseconds summed from
+ *     the statements' microsecond-resolution durations
  * @param sqlCount number of correlated SQL statements
  * @param sqlPercent percentage of the total request time spent in SQL, or {@code null}
  */
-public record RequestProfileTimingDto(Long totalMs, long sqlMs, int sqlCount, Double sqlPercent) {}
+public record RequestProfileTimingDto(Long totalMs, double sqlMs, int sqlCount, Double sqlPercent) {}
