+++ bootui-core/src/main/java/io/github/jdubois/bootui/core/dto/SqlAttributionBucketDto.java
@@ -9,13 +9,17 @@ package io.github.jdubois.bootui.core.dto;
  * silently distort every share and total on the panel.</p>
  *
  * @param executions retained executions in this bucket
- * @param totalDurationMillis summed duration of those executions
+ * @param totalDurationMillis summed duration of those executions, in fractional milliseconds
  * @param errorCount executions in this bucket that failed
  * @param shareOfRetainedTimePercent this bucket's share of the window's total database time, 0-100
  * @param reason plain-language explanation of why these executions are in this bucket
  */
 public record SqlAttributionBucketDto(
-        long executions, long totalDurationMillis, long errorCount, double shareOfRetainedTimePercent, String reason) {
+        long executions,
+        double totalDurationMillis,
+        long errorCount,
+        double shareOfRetainedTimePercent,
+        String reason) {
 
     public static SqlAttributionBucketDto empty(String reason) {
         return new SqlAttributionBucketDto(0, 0, 0, 0, reason);
